Refactor GpuExecutable helper responsibilities
