+++ xla/service/gpu/BUILD
@@ -695,6 +695,97 @@ bool_flag(
 )
 
 cc_library(
+    name = "gpu_executable_types",
+    srcs = ["gpu_executable_types.cc"],
+    hdrs = ["gpu_executable_types.h"],
+    deps = [
+        ":gpu_executable_proto_cc",
+        ":ir_emission_utils",
+        "//xla:literal",
+        "//xla:shape_tree",
+        "//xla/hlo/ir:hlo",
+        "//xla/service:buffer_assignment",
+        "//xla/service:hlo_proto_cc",
+        "//xla/stream_executor:device_address",
+        "@com_google_absl//absl/base:nullability",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/types:span",
+    ],
+)
+
+cc_library(
+    name = "gpu_executable_module_cache",
+    srcs = ["gpu_executable_module_cache.cc"],
+    hdrs = ["gpu_executable_module_cache.h"],
+    deps = [
+        ":gpu_executable_types",
+        "//xla:status_macros",
+        "//xla:util",
+        "//xla/stream_executor:device_address",
+        "//xla/stream_executor:module_spec",
+        "//xla/stream_executor:platform",
+        "//xla/stream_executor:scoped_module_handle",
+        "//xla/stream_executor:stream",
+        "//xla/stream_executor:stream_executor_h",
+        "//xla/stream_executor/cuda:cuda_platform_id",
+        "//xla/tsl/platform:status_macros",
+        "@com_google_absl//absl/base",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/log:check",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings:str_format",
+        "@com_google_absl//absl/synchronization",
+        "@com_google_absl//absl/types:span",
+    ],
+)
+
+cc_library(
+    name = "command_buffer_va_remapper",
+    srcs = ["command_buffer_va_remapper.cc"],
+    hdrs = ["command_buffer_va_remapper.h"],
+    deps = [
+        ":buffer_allocations",
+        "//xla:status_macros",
+        "//xla:util",
+        "//xla/backends/gpu/runtime:annotation",
+        "//xla/backends/gpu/runtime:command",
+        "//xla/backends/gpu/runtime:command_buffer_thunk",
+        "//xla/backends/gpu/runtime:thunk",
+        "//xla/backends/gpu/runtime:thunk_executor",
+        "//xla/runtime:buffer_use",
+        "//xla/service:buffer_assignment",
+        "//xla/service:service_executable_run_options",
+        "//xla/stream_executor:device_address",
+        "//xla/stream_executor:device_address_allocator",
+        "//xla/stream_executor:event",
+        "//xla/stream_executor:memory_allocation",
+        "//xla/stream_executor:memory_reservation",
+        "//xla/stream_executor:stream",
+        "//xla/stream_executor:stream_executor_h",
+        "//xla/stream_executor:vmm_device_address_allocator",
+        "//xla/tsl/platform:status_macros",
+        "//xla:xla_proto_cc",
+        "@com_google_absl//absl/base",
+        "@com_google_absl//absl/container:btree",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/container:node_hash_map",
+        "@com_google_absl//absl/functional:function_ref",
+        "@com_google_absl//absl/log",
+        "@com_google_absl//absl/log:check",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/strings:str_format",
+        "@com_google_absl//absl/strings:string_view",
+        "@com_google_absl//absl/synchronization",
+        "@com_google_absl//absl/types:span",
+        "@tsl//tsl/profiler/lib:scoped_annotation",
+    ],
+)
+
+cc_library(
     name = "gpu_executable",
     srcs = ["gpu_executable.cc"],
     hdrs = ["gpu_executable.h"],
@@ -702,9 +793,12 @@ cc_library(
         ":alias_info",
         ":backend_configs_cc",
         ":buffer_allocations",
+        ":command_buffer_va_remapper",
         ":gpu_constants",
+        ":gpu_executable_module_cache",
         ":gpu_executable_proto_cc",
         ":gpu_executable_run_options",
+        ":gpu_executable_types",
         ":ir_emission_utils",
         ":stream_executor_util",
         "//xla:executable_run_options",
@@ -726,9 +820,7 @@ cc_library(
         "//xla/backends/gpu/runtime:collective_memory_cache",
         "//xla/backends/gpu/runtime:collective_memory_requests",
         "//xla/backends/gpu/runtime:collective_params",
-        "//xla/backends/gpu/runtime:command",
         "//xla/backends/gpu/runtime:command_buffer_conversion_pass",
-        "//xla/backends/gpu/runtime:command_buffer_thunk",
         "//xla/backends/gpu/runtime:execution_stream_id",
         "//xla/backends/gpu/runtime:scratch_memory",
         "//xla/backends/gpu/runtime:scratch_memory_requests",
@@ -745,7 +837,6 @@ cc_library(
         "//xla/core/collectives:collectives_registry",
         "//xla/hlo/ir:hlo",
         "//xla/pjrt/proto:compile_options_proto_cc",
-        "//xla/runtime:buffer_use",
         "//xla/runtime:device_id",
         "//xla/runtime:hang_watchdog",
         "//xla/service:buffer_assignment",
@@ -769,12 +860,8 @@ cc_library(
         "//xla/stream_executor:event",
         "//xla/stream_executor:event_based_timer",
         "//xla/stream_executor:kernel_stats",
-        "//xla/stream_executor:memory_allocation",
-        "//xla/stream_executor:memory_reservation",
-        "//xla/stream_executor:module_spec",
         "//xla/stream_executor:platform",
         "//xla/stream_executor:platform_id",
-        "//xla/stream_executor:scoped_module_handle",
         "//xla/stream_executor:stream",
         "//xla/stream_executor:stream_executor_h",
         "//xla/stream_executor:vmm_device_address_allocator",
